requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
courlan==1.3.2
dateparser==1.2.2
ddgs==9.5.4
distro==1.9.0
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.31.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
narwhals==2.2.0
numpy==2.0.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
primp==0.15.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rpds-py==0.27.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.48.1
tenacity==9.1.2
tld==0.13.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
xlsxwriter==3.2.5
